pyproject.toml
[project]
name = "mlops-rk-project"
version = "0.1.0"
description = "Course project for MLOps midterm: e-commerce category classification"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
  "mlflow==3.4.0",
  "pandas==2.2.3",
  "numpy==2.1.3",
  "scikit-learn==1.5.2",
  "matplotlib==3.9.2",
  "PyYAML==6.0.2",
  "joblib==1.4.2",
  "requests==2.32.3",
  "beautifulsoup4==4.12.3",
  "lxml==5.3.0",
]

[project.optional-dependencies]
dev = [
  "pytest==8.3.3",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["project/src"]
